[avalanche] Fix the abc_p2p_proof_inventory test
Summary:
The test can fail under some condition because it waits for a proof to
become orphaned by waiting for the proof to be known by the node then
asserting its orphan state. But the proof is already known valid before
this call, so it is possible that the assertion fails because the proof
has not be oprhaned yet.
This diff fixes the test by waiting for the proof to turn orphan.

Ref T1611.

Test Plan:
  ninja check-functional

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Maniphest Tasks: T1611

Differential Revision: https://reviews.bitcoinabc.org/D9691