test: Add missing syncwithvalidationinterfacequeue
Summary:
The index on the block filters is running in the background on the validation interface. To avoid intermittent test failures, it needs to be synced.

This concludes backport of [[https://github.com/bitcoin/bitcoin/pull/22311 | core#22311]]
https://github.com/bitcoin/bitcoin/pull/22311/commits/fadddd13eef4428f5fa7237583d4be41a9335cd9

Depends on D11445

Test Plan: `test/functional/test_runner.py p2p_blockfilters`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11446