[avalanche] Move the pubkey storage from the node to the network layer
Summary:
This is not used by the peer manager, so no need to store it there. This
replaces the delegation in the CNode::AvalancheState which was unused.

Note that this diff introduces no locking mechanism to the avalanche
state which is now created/updated by the `addavalanchenode` RPC. This
might cause issues if the peer sends an avahello message at the same
time, which is expected to be a very unlikely event.

Ref T1634.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Maniphest Tasks: T1634

Differential Revision: https://reviews.bitcoinabc.org/D9719