[avalanche] Don't check the delegation in the peer manager
Summary:
The `PeerManager::addNode()` actually checks the delegation for the
node. This is node the right place to do the check since the peer
manager don't use the delegation ever but only the proof id, and also
doesn't return a proper error message in case of failure. This check is
already performed at the avahello reception which will be the main
callsite for this `addNode()` method.

Depends on D9719.

Ref T1634.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1634

Differential Revision: https://reviews.bitcoinabc.org/D9720