[avalanche] register local proof at first tip update
Summary:
Since we can now store proofs as orphans, it is possible to register the local proof before IBD is finished. When the blocks containing the stakes are known, the proof will be unorphaned.

Add a test for a node started with an orphan proof.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien, deadalnix

Reviewed By: #bitcoin_abc, Fabien, deadalnix

Subscribers: Fabien, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D9742