[avalanche] Use proof registration time instead of "proof time"
Summary:
The most puzzling about the existing code isn't that someone wrote it, it is that some read it and thought "yeah, this does make sense. We have compiler time, run time, and now proof time. Let's smash that accept button."

Change the type to use seconds, because this is what the network stack uses anyways and this is where it is consumed.

Test Plan: For some reason, this feature of the PeerManager is tested in the Processor test. So the test has been updated there. In any case the test suite passes.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9746