Remove the 1s timeouts in p2p_inv_download
Summary: This test fails occasionally on CI but I couldn't reproduce locally, so these timeouts might be too short when running the extended test suite. It's not a problem to wait for a minute (the default) when the test fails, and it doesn't change the test duration on success.

Test Plan:
  ./test/functional/test_runner.py p2p_inv_download

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D9801