make copyright headers more consistent
Summary:
"The Bitcoin developers" and "The Bitcoin Core developers" is the
dominant use. Fix the few occurences of "Developers" (capital D).

Remove one  occurence of a "-" between the authors and the year, and one extra "-" after the first year.

Replace "Bitcoin developers" with "The Bitcoin developers" in one file.

Test Plan:
Before:
```
$./contrib/devtools/copyright_header.py report .
34 with unexpected copyright holder names
```

After:
```
$./contrib/devtools/copyright_header.py report .
27 with unexpected copyright holder names
```

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9816