[doc] start a developer documentation for avalanche
Summary:
This is a stub documentation for the avalanche protocol, including the user facing
primitives (proofs, delegations...) and network messages.

The intended audience for this document is:
 - node developers
 - technically inclined users interested in the protocol

It does not details the proof format, as this is likely to change very soon.

Test Plan: Proof reading in a markdown preview tool.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9837