[LINTER] Handle errors from the Prettier linter
Summary: When the linter returns an error, it crashes arcanist (at best). This diff catches the error and return it as part of the linter output. Since the output can vary a lot there is no regex involved, just a raw dump a stderr.

Test Plan: Edit some .json file to make it seriously invalid (like adding a unexpected bracket), then run `arc lint` and check there is no crash and the error is reported.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9838