[Cashtab] add support for additional fiat currencies
Summary: After seeing a request for Vietnamese dong support, figured might as well add a few more based on traffic source for Cashtab users. I removed the order-of-magnitude sanity check on price as this was broken by Vietnamese Dong and also CoinGecko is keeping the BCHA api separate from the XEC one, so there is no longer the risk of a sudden switch.

Test Plan: Review new currencies to make sure no typos. Cycle through a few and try to break the app. There is a known issue where the UI does not fully lock on currency change, so if you quickly switch to another currency it may display improperly. A task is open for this and it will be addressed in another diff.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9849