[Cashtab] remove price api order of magnitude error check
Summary:
This error check was put in place in case Coingecko changed their endpoint to offer XEC price instead of BCHA. However, Coingecko is providing 2 separate endpoints, so that risk is much lower. Also, the entire app will move to the XEC endpoint soon.

The main purpose of this diff is to support more fiat currencies. This order of magnitude check always throws an error for currencies that are an order of magnitude off from USD, such as Vietnamese Dong.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9850