[Cashtab] Lock nav UI while new settings are loading
Summary:
T1701

Because Cashtab screens were developed incrementally, they have distinct `loading` context. This means that `Configure.js` will lock only the `Configure.js` portion of the screen when the app is in a `loading` state, and not the full UI including the bottom nav menu. This allows a user to navigate away from the `Configure.js` screen after a settings page and potentially see unloaded settings on another screen. For example, a user could change the currency on the `Configure.js` screen and then see the wrong fiat balance displayed on the homescreen because the app has not finished loading the new exchange rate.

This diff applies the global `loading` context to the full app for the `Wallet.js` and the `Configure.js` screens. The `Send.js` and `SendToken.js` screens use a different loading context and a separate refactor will be required to correct the issue there; a task is open for this.

Test Plan: `npm start`, navigate to `Configure.js`, change fiat currency, note that nav menu is locked until currency is loaded.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9869