[Cashtab] Lock full UI when utxo set is loading after send
Summary:
T1720

Right now the app is set to lock the UI on the Send and SendToken screens while the utxo set is refreshing. However, the user may still click off the Send or SendToken screen and then navigate back. This could allow the user to send a transaction with an invalid utxo set. The transaction won't actually send, they will get an error msg, but it's confusing and not great UX.

This diff locks the nav bar while the utxo set refreshes.

Test Plan: `npm start`, send an XEC transaction, notice that the nav bar is locked, send an eToken transaction, notice the nav bar is locked

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9872