[net processing] Comment all TxRelay members
Summary:
This fully comments all the TxRelay members. The only significant change
is to the comment for m_relay_txs. Previously the comment stated that
one of the purposes of the field was that "We don't relay tx invs before
receiving the peer's version message". However, even without the
m_relay_txs flag, we would not send transactions to the peer before
receiving the `version` message, since SendMessages() returns
immediately if fSuccessfullyConnected is not set to true, which only
happens once a `version` and `verack` message have been received.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/22778 | core#22778]]
https://github.com/bitcoin/bitcoin/pull/22778/commits/290a8dab0288344fa5731ec2ffd09478e9420a2f

Depends on D14738

Test Plan: proof-reading

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D14739