[Cashtab] Remove support for sending to BCH or SLP addresses
Summary:
T1735

Remove support for sending to addresses with `bitcoincash:` or `simpleledger:` prefixes, even if they are valid cashaddr.

This diff also modifies validation so that xec must be sent to an `ecash:` prefixed address and eTokens must be sent to an `etoken:` prefixed address.

Test Plan: `npm start`, try to send XEC to a `bitcoincash:` prefixed address and an `etoken:` prefixed address and note validation, try to send eToken to `simpleledger:` prefixed address and `ecash:` prefixed address and note validation

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9883