[Cashtab] Remove unused artifact params from Send.js
Summary:
T1724

The `Send.js` screen accepts two props --- `filledAddress` and `callbackTxid` --- that are never used in Cashtab. These are artifacts of the original Mint codebase where the `Send` component was a pop-up that could be called in different contexts, e.g. to send dividend payments.

Cashtab has diverged so much from Mint at this point that even if similar functionality were to be built back in, it would not use this same approach. Therefore they should be removed as they are baffling to anyone without a familiarity with the entire history of the codebase.

Test Plan: `npm start` and send an XEC transaction

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9884