[Cashtab] Handle unconfirmed genesis tx
Summary:
T1752

Cashtab displays `12/31/1969` for `Genesis Date` on the `SendToken.js` page when a token's genesis transaction has not yet confirmed. This is because `tokenStats.timestampUnix` is returned as null in this case.

Handle this case and display an informational note to let the user know the token was not created in the 60s.

Test Plan:
`npm start`
Create a token
Click the token tile to send some of your newly created token
Observe `Genesis Date` reads `Just now (Genesis tx confirming)` if tx is unconfirmed

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D9892