[Cashtab] Fix bug where new user sees permanent loading screen
Summary:
T1752

In standardizing the wallet loading screeen to lock the entire app UI instead of screen by screen, I neglected the case of a new user with no created wallet.

This user will see a perpetual loading screen since `wallet === false` and therefore `validWallet` will always be false.

This diff handles this case in the App.js spinner.

Test Plan:
`npm start`
Load in an incognito window or browser with no cashtab wallet
Observe that the app does load
Create a wallet
Send a tx and verify the app locks UI for loading and then unlocks when utxo set is refreshed

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9893