[Cashtab] Deprecate obsolete websocket support
Summary:
T1762

Cashtab supported websockets for incoming txs when it launched, because blockchain.com had a compatible BCH websocket.

However, this is ninsight-based and does not support current eCash infra. Moreover, the websocket part of the codebase is obsolete in various ways as it was not maintained in line with the app.

This diff removes this obsolete websocket code.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9910