[init] Read/decode asmap before constructing addrman
Summary:
```
Commit 181a1207 introduced an initialization order bug: CAddrMan's m_asmap must be set before deserializing peers.dat. Restore that ordering.
```

Note that we didn't introduce the bug (yet), mostly because most of the referenced PR is not relevant for us so it has been purposedly delayed. A note is added to the code so that later backport will not introduce more issues.

Partial backport of [[https://github.com/bitcoin/bitcoin/pull/22791 | core#22791]]:
https://github.com/bitcoin/bitcoin/pull/22791/commits/50fd77045e2f858a53486b5e02e1798c92ab946c

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12291