[Cashtab] Patch bug in address display
Summary:
T1823

I introduced a bug when I reconfigured the address display. Cashtab will copy the correct address if you click the QR code or the address, however it is displaying a repeated string in the actual address. Users who manually select that address will get the wrong thing.

This diff corrects the oversight.

Test Plan: `npm start`, click the QR code to copy the address and paste it somewhere, manually highlight the address and copy paste it, verify they are the same; do the same with etoken

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10044