[Cashtab] correct error in parsing tx history for op ret txs
Summary:
T1847

Cashtab assumes that a tx with no address object in output is a token tx when parsing tx history. However, the tx could be a normal OP return tx.

This diff flags such a tx as non-token if no token info is available in the API check. Previous behavior failed on the error.

This does not impact utxo parsing, only tx history parsing.

Test Plan:
1. DM me for a seed with such a tx in history to test, or send yourself a tx with an OP return output.
2. Observe that the tx is parsed correctly.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10102