[avalanche] Add a payout script to the proof
Summary:
This will be used for stake rewards. Since it's changing the proof format, I hid it behind a flag so I don't have to update all the tests, especially since more format changes are coming (and updating the hardcoded proofs in tests is expensive).

The payout script cannot be changed without signing the stakes again. This will change in the future.

Ref T1676.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1676

Differential Revision: https://reviews.bitcoinabc.org/D10109