Make all CRPCCommand actors take a const Config &
Summary:
After D9745 there seems to no longer be a need for RPC commands that
take a non-const Config as a parameter.
In D10007 and D10010, it seemed that the `help` RPC still needed a
non-const `Config &`, but it was actually no longer the case.
The `help` executes the other RPCs to get their help message, and since
D9745 removed `setexcessiveblock`, this is no longer needed.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10149