validation: LoadChainstate only needs Consensus::Params
Summary:
We used to pass Config before the verification was split.

The original commit is about changing the signature of VerifyDB, but this is not applicable to our codebase because we need the config to get the max block size.

Partial backport of [[https://github.com/bitcoin/bitcoin/pull/23280 | core#23280]]:
https://github.com/bitcoin/bitcoin/pull/23280/commits/15f2e33bb3d1ad3bc997f6a84956337f46495091

Depends on D12578.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12579